+++ docs/sphinx/changelog.rst
@@ -18,6 +18,7 @@ New Features
 Internal Changes
 ================
 - Generalized the size of dRdF (:pull:`56`). by `Nathan Miller`_.
+- Added the initialization of the unknown vector (:pull:`60`). By `Nathan Miller`_.
 
 Bug Fixes
 =========